syncer: split shared package
